backend/requirements.txt
# FastAPI & Server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
websockets==12.0

# Database
sqlmodel==0.0.14
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# AI
openai==1.12.0
tiktoken==0.6.0

# Celery
celery[redis]==5.3.6
redis==5.0.1

# File Processing
PyPDF2==3.0.1
python-docx==1.1.0

# Email & SMS
sendgrid==6.11.0
twilio==8.13.0

# Reports
reportlab==4.1.0
openpyxl==3.1.2

# Templates
jinja2==3.1.3

# Settings
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP Client
httpx==0.26.0

# Utilities
python-dateutil==2.8.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.4

# Linting
ruff==0.2.1
